+++ contracts/Swap/IDexManager.sol
@@ -21,6 +21,6 @@ interface IDexManager {
         address _account,
         uint256 _amountIn,
         uint256 _deadline,
-        uint8 _slippage
+        uint16 _slippage
     ) external;
 }
